requirements.txt
# SENTINEL-AI Dependencies
# Core Framework Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# LLM and Embeddings (Ollama)
langchain-ollama>=0.1.0

# Vector Database
chromadb>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Web Framework
streamlit>=1.35.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security and Validation
validators>=0.20.0
regex>=2023.8.8

# HTTP Requests and Data Fetching
requests>=2.31.0
httpx>=0.24.0

# STIX/TAXII for MITRE ATT&CK
stix2>=3.0.0

# Additional Utilities
tqdm>=4.65.0
loguru>=0.7.0
click>=8.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0